Rename form modules
