--- import_export_extensions/admin/forms/base_import_form.py
@@ -1,11 +1,0 @@
-from django import forms
-
-from import_export import forms as base_forms
-
-
-class ExtendedImportForm(base_forms.ImportForm):
-    """Import form with `force_import` option."""
-    force_import = forms.BooleanField(
-        required=False,
-        initial=False,
-    )
